Improve is_power_of_two()
