+++ ethereum/datastructures/src/test/java/tech/pegasys/artemis/datastructures/util/BeaconStateUtilTest.java
@@ -13,10 +13,12 @@
 
 package tech.pegasys.artemis.datastructures.util;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static tech.pegasys.artemis.datastructures.util.BeaconStateUtil.is_power_of_two;
 import static tech.pegasys.artemis.datastructures.util.DataStructureUtil.randomDepositInput;
 import static tech.pegasys.artemis.datastructures.util.DataStructureUtil.randomUnsignedLong;
 import static tech.pegasys.artemis.datastructures.util.DataStructureUtil.randomValidator;
@@ -329,6 +331,20 @@ class BeaconStateUtilTest {
     assertEquals(expectedBadActorBalance, beaconState.getValidator_balances().get(validatorIndex));
     assertEquals(
         expectedWhistleblowerBalance, beaconState.getValidator_balances().get(whistleblowerIndex));
+  }
+
+  @Test
+  void isPowerOfTwo() {
+    // Not powers of two:
+    assertThat(is_power_of_two(UnsignedLong.ZERO)).isEqualTo(false);
+    assertThat(is_power_of_two(UnsignedLong.valueOf(42L))).isEqualTo(false);
+    assertThat(is_power_of_two(UnsignedLong.valueOf(Long.MAX_VALUE))).isEqualTo(false);
+    // Powers of two:
+    assertThat(is_power_of_two(UnsignedLong.ONE)).isEqualTo(true);
+    assertThat(is_power_of_two(UnsignedLong.ONE.plus(UnsignedLong.ONE))).isEqualTo(true);
+    assertThat(is_power_of_two(UnsignedLong.valueOf(0x040000L))).isEqualTo(true);
+    assertThat(is_power_of_two(UnsignedLong.valueOf(0x0100000000L))).isEqualTo(true);
+    assertThat(is_power_of_two(UnsignedLong.fromLongBits(0x8000000000000000L))).isEqualTo(true);
   }
 
   private BeaconState createBeaconState() {
